Router using authContext
Fixes #136
