+++ src/App.js
@@ -1,46 +1,19 @@
-import React from 'react'
-import {
-  Routes,
-  Route,
-  useNavigate
-} from "react-router-dom";
+import { RouterProvider } from 'react-router-dom';
 import { NextUIProvider } from "@nextui-org/react";
 
-import { Layout} from '@douyinfe/semi-ui';
+import AuthContextProvider from "./Contexts/authContext";
+
+import { router } from "./Config/Router/BrowserRouter";
 import './Styles/App.scss';
-import Navbar from './Components/Navbar'
-import FooterCustom from './Components/footer.js';
-import Landing from './Pages/Landing';
-import QuieroInvertir from './Pages/QuieroInvertir';
-import SolicitarPrestamo from './Pages/SolicitarPrestamo';
-import CreaTuCuenta from './Pages/SignUp/CreaTuCuenta';
-import IngresaTusDatos from './Pages/SignUp/IngresaTusDatos';
-import VerificacionCorreo from './Pages/SignUp/VerificacionCorreo';
-import VerificacionTelefono from './Pages/SignUp/VerificacionTelefono';
 
 function App() {
-  const { Header, Sider, Footer, Content } = Layout;
-  const navigate = useNavigate();
-
   return (
-    <NextUIProvider navigate={navigate}>
-        <Layout>
-          <Navbar/>
-          <Content className='app-content'>
-            <Routes>
-              <Route path='/' element = { <Landing/> } />
-              <Route path='/quiero-invertir' element = { <QuieroInvertir/> } />
-              <Route path='/solicitar-prestamo' element = { <SolicitarPrestamo/> } />
-              <Route path='/registro' element = { <CreaTuCuenta/> } />
-              <Route path='/ingresar-datos/:id_usuario' element={<IngresaTusDatos />} />
-              <Route path='/verificar-correo/:id_usuario' element={<VerificacionCorreo />} />
-              <Route path='/verificar-telefono/:id_usuario' element={<VerificacionTelefono />} />
-            </Routes>
-          </Content>
-          <Footer>
-            <FooterCustom />
-          </Footer>
-        </Layout>
+    <NextUIProvider>
+      <div className="App">
+        <AuthContextProvider>
+          <RouterProvider router={router} />
+        </AuthContextProvider>
+      </div>
     </NextUIProvider>
   );
 }
