+++ src/Config/Router/paths.js
@@ -1,0 +1,14 @@
+export const LOGIN = '/login';
+export const SIGNUP = '/signup';
+export const PRIVATE = '/private';
+
+export const LOGOUT = '/private/logout';
+
+export const QUIERO_INVERTIR = '/quiero-invertir';
+export const VERIFICAR = '/verificar';
+export const INGRESAR_DATOS = '/verificar/ingresar-datos';
+export const VERIFICAR_CORREO = '/verificar/verificar-correo';
+export const VERIFICAR_TELEFONO = '/verificar/verificar-telefono';
+
+export const SOLICITAR_PRESTAMO = '/solicitar-prestamo';
+export const TU_HISTORIAL = '/solicitar-prestamo/tu-historial';
