+++ src/Pages/Login.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Login() {
+  return (
+    <div>Login</div>
+  )
+}
+
+export default Login
