--- src/Pages/SolicitarPrestamo.js
@@ -1,13 +1,0 @@
-import React from 'react'
-
-import FormDatosUsuario from '../Components/SolicitarPrestamo/FormDatosUsuario'
-
-function SolicitarPrestamo() {
-  return (
-    <div className=''>
-      <FormDatosUsuario />
-    </div>
-  )
-}
-
-export default SolicitarPrestamo
